ntools_fispact

Struct Inventory

pub struct Inventory {
    pub run_data: RunData,
    pub intervals: Vec<Interval>,
}

Top level inventory data

Collection of the run metadata and list of results for each FISPACT-II calculation step.

Useful functionality is implemented at this level for investigating properties across intervals, and for collecting broader summaries of data.

Several tweaks are made during deserialisation, so data structures are not an exact 1:1 match for the JSON dictionary.

Changes to FISPACT-II structure

All masses are converted to grams in the background such that Nuclide and Interval mass values are consistent.

Sample dose rate for intervals are converted into a more concise structure. See Dose and DoseKind for details.

Several of the “totals” are renamed for brevity. e.g. “total_activity” is simply stored as Interval.activity.

Fields
run_data: RunData

Metadata for the run

intervals: Vec<Interval>

Collection of all run intervals

Implementations

impl Inventory

pub fn from_json<P: AsRef<Path>>(path: P) -> Result<Inventory, Error>

Read the full JSON inventory results

Returns the full Inventory data, with all Intervals and metadata.

path is anything that can be turned into a Path, such as a &str, String, Path, etc…

Example:

// Read the JSON inventory data
let inventory = Inventory::from_json("path/to/results.json");

pub fn activity_list(&self) -> Vec<f64>

List of total activity (Bq) for each Interval

pub fn specific_activity_list(&self) -> Vec<f64>

List of sample specific activity (Bq/g) for each Interval

pub fn dose_list(&self) -> Vec<Dose>

Collection of total Dose rates for each Interval

pub fn mass_list(&self) -> Vec<f64>

List of total sample mass (g) for each Interval

pub fn total_times(&self) -> Vec<f64>

List of total times (seconds) for each Interval

pub fn irradiation_times(&self) -> Vec<f64>

List of irradiation times (seconds) for each Interval

pub fn cooling_times(&self) -> Vec<f64>

List of cooling times (seconds) for each Interval

pub fn nuclides(&self) -> Vec<Nuclide>

List of any matching Nuclide objects throughout the Inventory

pub fn element_names(&self) -> Vec<String>

List of names for all unique elements in the Inventory

pub fn nuclide_names(&self) -> Vec<String>

List of names for all unique nuclides in the Inventory

pub fn nuclide_transient(&self) -> Vec<f64>

List of data for some time dependednt transient

pub fn nearest_interval(&self) -> &Interval

Finds the nearest Interval by total time

pub fn normalise_flux(&mut self, _factor: f64)

Applies a flux normalisation to all data in the Inventory